LPD-7822 LPD-8351 portal-search-spi: Move contributors to new package
https://liferay.atlassian.net/browse/LPD-8351

# breaking

## What modules/apps/portal-search/portal-search-spi/src/main/java/com/liferay/portal/search/spi/model/permission/SearchPermissionFilterContributor.java

modules/apps/portal-search/portal-search-spi/src/main/java/com/liferay/portal/search/spi/model/permission/SearchPermissionFilterContributor.java was moved to
modules/apps/portal-search/portal-search-spi/src/main/java/com/liferay/portal/search/spi/model/permission/contributor/SearchPermissionFilterContributor.java

## Why

Most contributors are located in contributor packages

----
